+++ config.example.ts
@@ -1,23 +1,26 @@
+import UserAgent = require('user-agents');
+const Agent = new UserAgent().random().data;
+
 /**
- * X-Super-Properties header for Discord. This **can be** filled with random values (note: *NOT KEYS*); it has no effect on functionality.
+ * X-Super-Properties header for Discord. Do **NOT** modify this unless you know what you're doing!
  */
 const super_properties = Buffer.from(JSON.stringify({
-    'os': 'Windows',
-    'browser': 'Firefox',
+    'os': Agent.platform,
+    'browser': ['Firefox', 'Chrome', 'Edge', 'Brave'][Math.floor(Math.random() * 4)],
     'device':'',
-    'browser_user_agent':'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:72.0) Gecko/20100101 Firefox/72.0',
-    'browser_version':'72.0',
-    'os_version':'10',
-    'referrer':'',
-    'referring_domain':'',
-    'referrer_current':'',
-    'referring_domain_current':'',
-    'release_channel':'stable',
-    'client_build_number':51863,
-    'client_event_source':null}
-)).toString('base64');
+    'browser_user_agent': Agent.userAgent,
+    'browser_version': (Math.random() * (999.99 - 1.0 + 1) + 1.0).toFixed(3),
+    'os_version': '' + [10, 9, 8, 7][Math.floor(Math.random() * 4)],
+    'referrer': '',
+    'referring_domain': '',
+    'referrer_current': '',
+    'referring_domain_current': '',
+    'release_channel': 'stable',
+    'client_build_number': Math.floor(Math.random() * (99999 - 10000 + 1) + 10000),
+    'client_event_source': null
+})).toString('base64');
 
-const useragent = 'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:72.0) Gecko/20100101 Firefox/72.0';
+const useragent = Agent.userAgent;
 const smspva = '';
 const captcha = '';